# nodejs/api/swagger.yaml
openapi: 3.0.0
info:
  title: Unified Data Repository Service API file for subscription data
  description: |
    Unified Data Repository Service (subscription data).
    The API version is defined in 3GPP TS 29.504.
    © 2019, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC).
    All rights reserved.
  version: '-'
...
servers:
- url: /
paths:
...
  /subscription-data/{ueId}/{servingPlmnId}/provisioned-data:
    get:
      tags:
      - Provisioned Data (Document)
      summary: Retrieve multiple provisioned data sets of a UE
      operationId: QueryProvisionedData
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: servingPlmnId
        in: path
        description: PLMN ID
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarPlmnId'
      - name: dataset-names
        in: query
        description: List of dataset names
        required: false
        style: form
        explode: false
        schema:
          $ref: '#/components/schemas/DatasetNames'
      responses:
        200:
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProvisionedDataSets'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: ProvisionedDataDocument
...
  /subscription-data/{ueId}/context-data/smf-registrations/{pduSessionId}:
    get:
      tags:
      - SMF Registration (Document)
      summary: Retrieves the individual SMF registration of a UE
      operationId: QuerySmfRegistration
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: pduSessionId
        in: path
        description: PDU session id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/PduSessionId'
      - name: fields
        in: query
        description: attributes to be retrieved
        required: false
        style: form
        explode: false
        schema:
          minItems: 1
          type: array
          items:
            type: string
      - name: supported-features
        in: query
        description: Supported Features
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/SupportedFeatures'
      responses:
        200:
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmfRegistration'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: SMFRegistrationDocument
    put:
      tags:
      - SMF Registration (Document)
      summary: To create an individual SMF context data of a UE in the UDR
      operationId: CreateSmfContextNon3gpp
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: pduSessionId
        in: path
        description: PDU session id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/PduSessionId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmfRegistration'
      responses:
        201:
          description: Upon success, a response body containing a representation of
            the created Individual SmfRegistration resource shall be returned
        default:
          description: Unexpected error
          headers:
            Location:
              description: 'Contains the URI of the newly created resource, according
                to the structure: {apiRoot}/nudr-dr/<apiVersion>/subscription-data/{ueId}/context-data/smf-registrations/{pduSessionId}'
              required: true
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: SMFRegistrationDocument
    delete:
      tags:
      - SMF Registration (Document)
      summary: To remove an individual SMF context data of a UE the UDR
      operationId: DeleteSmfContext
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: pduSessionId
        in: path
        description: PDU session id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/PduSessionId'
      responses:
        204:
          description: Upon success, an empty response body shall be returned.
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: SMFRegistrationDocument
...
  /subscription-data/{ueId}/pp-data:
    get:
      tags:
      - Parameter Provision (Document)
      summary: Read the profile of a given UE
      operationId: GetppData
      parameters:
      - name: ueId
        in: path
        description: pp data for a UE
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: supported-features
        in: query
        description: Supported Features
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/SupportedFeatures'
      - name: If-None-Match
        in: header
        description: Validator for conditional requests, as described in RFC 7232,
          3.2
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: If-Modified-Since
        in: header
        description: Validator for conditional requests, as described in RFC 7232,
          3.3
        required: false
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: Expected response to a valid request
          headers:
            Cache-Control:
              description: Cache-Control containing max-age, as described in RFC 7234,
                5.2
              style: simple
              explode: false
              schema:
                type: string
            ETag:
              description: Entity Tag, containing a strong validator, as described
                in RFC 7232, 2.3
              style: simple
              explode: false
              schema:
                type: string
            Last-Modified:
              description: Timestamp for last modification of the resource, as described
                in RFC 7232, 2.2
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PpData'
        403:
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: ParameterProvisionDocument
    patch:
      tags:
      - ProvisionedParameterData (Document)
      summary: modify the provisioned parameter data
      operationId: ModifyPpData
      parameters:
      - name: ueId
        in: path
        description: pp data for a UE
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      requestBody:
        content:
          application/json-patch+json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PatchItem'
        required: true
      responses:
        204:
          description: Expected response to a valid request
        403:
          description: modification is rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: ProvisionedParameterDataDocument
...
  /subscription-data/{ueId}/context-data/sdm-subscriptions:
    get:
      tags:
      - SDM Subscriptions (Collection)
      summary: Retrieves the sdm subscriptions of a UE
      operationId: Querysdmsubscriptions
      parameters:
      - name: ueId
        in: path
        description: UE id
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: supported-features
        in: query
        description: Supported Features
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/SupportedFeatures'
      responses:
        200:
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SdmSubscription'
                x-content-type: application/json
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: SDMSubscriptionsCollection
    post:
      tags:
      - SDM Subscriptions (Collection)
      summary: Create individual sdm subscription
      operationId: CreateSdmSubscriptions
      parameters:
      - name: ueId
        in: path
        description: UE ID
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SdmSubscription'
        required: true
      responses:
        201:
          description: Expected response to a valid request
          headers:
            Location:
              description: 'Contains the URI of the newly created resource, according
                to the structure: {apiRoot}/nudr-dr/<apiVersion>/subscription-data/{ueId}/context-data/sdm-subscriptions/{subsId}'
              required: true
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SdmSubscription'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: SDMSubscriptionsCollection
...
  /subscription-data/{ueId}/identity-data:
    get:
      tags:
      - Query Identity Data by SUPI or GPSI (Document)
      summary: Retrieve identity data by SUPI or GPSI
      operationId: GetIdentityData
      parameters:
      - name: ueId
        in: path
        description: UE ID
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/VarUeId'
      - name: If-None-Match
        in: header
        description: Validator for conditional requests, as described in RFC 7232,
          3.2
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: If-Modified-Since
        in: header
        description: Validator for conditional requests, as described in RFC 7232,
          3.3
        required: false
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: OK
          headers:
            Cache-Control:
              description: Cache-Control containing max-age, as described in RFC 7234,
                5.2
              style: simple
              explode: false
              schema:
                type: string
            ETag:
              description: Entity Tag, containing a strong validator, as described
                in RFC 7232, 2.3
              style: simple
              explode: false
              schema:
                type: string
            Last-Modified:
              description: Timestamp for last modification of the resource, as described
                in RFC 7232, 2.2
              style: simple
              explode: false
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityData'
        default:
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      x-swagger-router-controller: QueryIdentityDataBySUPIOrGPSIDocument
...
components:
  schemas:
...
    VarPlmnId:
      pattern: ^[0-9]{5,6}$
      type: string
    DatasetNames:
      minItems: 1
      uniqueItems: true
      type: array
      items:
        $ref: '#/components/schemas/DataSetName'
    DataSetName:
      anyOf:
      - type: string
        enum:
        - AM
        - SMF_SEL
        - SMS_SUB
        - SM
        - TRACE
        - SMS_MNG
      - type: string
    ProvisionedDataSets:
      type: object
      properties:
        amData:
          $ref: '#/components/schemas/AccessAndMobilitySubscriptionData'
        smfSelData:
          $ref: '#/components/schemas/SmfSelectionSubscriptionData'
        smsSubsData:
          $ref: '#/components/schemas/SmsSubscriptionData'
        smData:
          type: array
          items:
            $ref: '#/components/schemas/SessionManagementSubscriptionData'
        traceData:
          $ref: '#/components/schemas/TraceData'
        smsMngData:
          $ref: '#/components/schemas/SmsManagementSubscriptionData'
    AccessAndMobilitySubscriptionData:
      type: object
      properties:
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
        gpsis:
          type: array
          items:
            $ref: '#/components/schemas/Gpsi'
        internalGroupIds:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/GroupId'
        subscribedUeAmbr:
          $ref: '#/components/schemas/AmbrRm'
        nssai:
          $ref: '#/components/schemas/Nssai'
        ratRestrictions:
          type: array
          items:
            $ref: '#/components/schemas/RatType'
        forbiddenAreas:
          type: array
          items:
            $ref: '#/components/schemas/Area'
        serviceAreaRestriction:
          $ref: '#/components/schemas/ServiceAreaRestriction'
        coreNetworkTypeRestrictions:
          type: array
          items:
            $ref: '#/components/schemas/CoreNetworkType'
        rfspIndex:
          $ref: '#/components/schemas/RfspIndexRm'
        subsRegTimer:
          $ref: '#/components/schemas/DurationSecRm'
        ueUsageType:
          $ref: '#/components/schemas/UeUsageType'
        mpsPriority:
          $ref: '#/components/schemas/MpsPriorityIndicator'
        mcsPriority:
          $ref: '#/components/schemas/McsPriorityIndicator'
        activeTime:
          $ref: '#/components/schemas/DurationSecRm'
        dlPacketCount:
          $ref: '#/components/schemas/DlPacketCount'
        sorInfo:
          $ref: '#/components/schemas/SorInfo'
        upuInfo:
          $ref: '#/components/schemas/UpuInfo'
        micoAllowed:
          $ref: '#/components/schemas/MicoAllowed'
        sharedAmDataIds:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/SharedDataId'
        odbPacketServices:
          $ref: '#/components/schemas/OdbPacketServices'
        subscribedDnnList:
          type: array
          items:
            $ref: '#/components/schemas/Dnn'
    SmfSelectionSubscriptionData:
      type: object
      properties:
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
        subscribedSnssaiInfos:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SnssaiInfo'
        sharedSnssaiInfosId:
          $ref: '#/components/schemas/SharedDataId'
...
    Dnn:
      type: string
    SessionManagementSubscriptionData:
      required:
      - singleNssai
      type: object
      properties:
        singleNssai:
          $ref: '#/components/schemas/Snssai'
        dnnConfigurations:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/DnnConfiguration'
          description: A map (list of key-value pairs where Dnn serves as key) of
            DnnConfigurations
        internalGroupIds:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/GroupId'
        sharedDnnConfigurationsId:
          $ref: '#/components/schemas/SharedDataId'
        odbPacketServices:
          $ref: '#/components/schemas/OdbPacketServices'
...
    SmfRegistration:
      required:
      - pduSessionId
      - plmnId
      - singleNssai
      - smfInstanceId
      type: object
      properties:
        smfInstanceId:
          $ref: '#/components/schemas/NfInstanceId'
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
        pduSessionId:
          $ref: '#/components/schemas/PduSessionId'
        singleNssai:
          $ref: '#/components/schemas/Snssai'
        dnn:
          $ref: '#/components/schemas/Dnn'
        emergencyServices:
          type: boolean
        pcscfRestorationCallbackUri:
          $ref: '#/components/schemas/Uri'
        plmnId:
          $ref: '#/components/schemas/PlmnId'
        pgwFqdn:
          type: string
...
    SmsManagementSubscriptionData:
      type: object
      properties:
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
        mtSmsSubscribed:
          type: boolean
        mtSmsBarringAll:
          type: boolean
        mtSmsBarringRoaming:
          type: boolean
        moSmsSubscribed:
          type: boolean
        moSmsBarringAll:
          type: boolean
        moSmsBarringRoaming:
          type: boolean
        sharedSmsMngDataIds:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/SharedDataId'
    SmsSubscriptionData:
      type: object
      properties:
        smsSubscribed:
          $ref: '#/components/schemas/SmsSubscribed'
        sharedSmsSubsDataId:
          $ref: '#/components/schemas/SharedDataId'
...
    PpData:
      type: object
      properties:
        communicationCharacteristics:
          $ref: '#/components/schemas/CommunicationCharacteristics'
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
...
    SdmSubscription:
      required:
      - callbackReference
      - monitoredResourceUris
      - nfInstanceId
      type: object
      properties:
        nfInstanceId:
          $ref: '#/components/schemas/NfInstanceId'
        implicitUnsubscribe:
          type: boolean
        expires:
          $ref: '#/components/schemas/DateTime'
        callbackReference:
          $ref: '#/components/schemas/Uri'
        amfServiceName:
          $ref: '#/components/schemas/ServiceName'
        monitoredResourceUris:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Uri'
        singleNssai:
          $ref: '#/components/schemas/Snssai'
        dnn:
          $ref: '#/components/schemas/Dnn'
        subscriptionId:
          type: string
        plmnId:
          $ref: '#/components/schemas/PlmnId'
...
    IdentityData:
      type: object
      properties:
        supiList:
          maxItems: 2
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Supi'
        gpsiList:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Gpsi'
...
    VarUeId:
      pattern: ^(imsi-[0-9]{5,15}|nai-.+|msisdn-[0-9]{5,15}|extid-[^@]+@[^@]+|.+)$
      type: string
    SupportedFeatures:
      pattern: ^[A-Fa-f0-9]*$
      type: string
    ProblemDetails:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/Uri'
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          $ref: '#/components/schemas/Uri'
        cause:
          type: string
        invalidParams:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/InvalidParam'
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
    Uri:
      type: string
    InvalidParam:
      required:
      - param
      type: object
      properties:
        param:
          type: string
        reason:
          type: string
    PatchItem:
      required:
      - op
      - path
      type: object
      properties:
        op:
          $ref: '#/components/schemas/PatchOperation'
        path:
          type: string
        from:
          type: string
        value:
          nullable: true
    PatchOperation:
      anyOf:
      - type: string
        enum:
        - add
        - copy
        - move
        - remove
        - replace
        - test
      - type: string
...
    PduSessionId:
      maximum: 255
      minimum: 0
      type: integer
...
    SharedDataId:
      pattern: ^[0-9]{5,6}-.+$
      type: string
...
    Nssai:
      required:
      - defaultSingleNssais
      type: object
      properties:
        supportedFeatures:
          $ref: '#/components/schemas/SupportedFeatures'
        defaultSingleNssais:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Snssai'
        singleNssais:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Snssai'
      nullable: true
    UeUsageType:
      type: integer
    MpsPriorityIndicator:
      type: boolean
    McsPriorityIndicator:
      type: boolean
    DlPacketCount:
      minimum: -1
      type: integer
    SorInfo:
      required:
      - ackInd
      - provisioningTime
      type: object
      properties:
        steeringContainer:
          $ref: '#/components/schemas/SteeringContainer'
        ackInd:
          $ref: '#/components/schemas/AckInd'
        sorMacIausf:
          $ref: '#/components/schemas/SorMac'
        countersor:
          $ref: '#/components/schemas/CounterSor'
        provisioningTime:
          $ref: '#/components/schemas/DateTime'
...
    UpuInfo:
      required:
      - provisioningTime
      - upuAckInd
      - upuDataList
      - upuRegInd
      type: object
      properties:
        upuDataList:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/UpuData_2'
        upuRegInd:
          $ref: '#/components/schemas/UpuRegInd'
        upuAckInd:
          $ref: '#/components/schemas/UpuAckInd'
        upuMacIausf:
          $ref: '#/components/schemas/UpuMac'
        counterUpu:
          $ref: '#/components/schemas/CounterUpu'
        provisioningTime:
          $ref: '#/components/schemas/DateTime'
...
    MicoAllowed:
      type: boolean
    SmsSubscribed:
      type: boolean
...
    SnssaiInfo:
      required:
      - dnnInfos
      type: object
      properties:
        dnnInfos:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/DnnInfo'
...
    NfInstanceId:
      type: string
      format: uuid
    TraceData:
      required:
      - eventList
      - neTypeList
      - traceDepth
      - traceRef
      type: object
      properties:
        traceRef:
          pattern: ^[0-9]{3}[0-9]{2,3}-[A-Fa-f0-9]{6}$
          type: string
        traceDepth:
          $ref: '#/components/schemas/TraceDepth'
        neTypeList:
          pattern: ^[A-Fa-f0-9]+$
          type: string
        eventList:
          pattern: ^[A-Fa-f0-9]+$
          type: string
        collectionEntityIpv4Addr:
          $ref: '#/components/schemas/Ipv4Addr'
        collectionEntityIpv6Addr:
          $ref: '#/components/schemas/Ipv6Addr'
        interfaceList:
          pattern: ^[A-Fa-f0-9]+$
          type: string
      nullable: true
    TraceDepth:
      anyOf:
      - type: string
        enum:
        - MINIMUM
        - MEDIUM
        - MAXIMUM
        - MINIMUM_WO_VENDOR_EXTENSION
        - MEDIUM_WO_VENDOR_EXTENSION
        - MAXIMUM_WO_VENDOR_EXTENSION
      - type: string
    Ipv4Addr:
      pattern: ^(([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])$
      type: string
      example: 198.51.100.1
    Ipv6Addr:
      type: string
      example: 2001:db8:85a3::8a2e:370:7334
      allOf:
      - pattern: ^((:|(0?|([1-9a-f][0-9a-f]{0,3}))):)((0?|([1-9a-f][0-9a-f]{0,3})):){0,6}(:|(0?|([1-9a-f][0-9a-f]{0,3})))$
      - pattern: ^((([^:]+:){7}([^:]+))|((([^:]+:)*[^:]+)?::(([^:]+:)*[^:]+)?))$
...
    GroupId:
      pattern: ^[A-Fa-f0-9]{8}-[0-9]{3}-[0-9]{2,3}-([A-Fa-f0-9][A-Fa-f0-9]){1,10}$
      type: string
...
    Gpsi:
      pattern: ^(msisdn-[0-9]{5,15}|extid-[^@]+@[^@]+|.+)$
      type: string
    AmbrRm:
      required:
      - downlink
      - uplink
      type: object
      properties:
        uplink:
          $ref: '#/components/schemas/BitRate'
        downlink:
          $ref: '#/components/schemas/BitRate'
      nullable: true
...
    RatType:
      anyOf:
      - type: string
        enum:
        - NR
        - EUTRA
        - WLAN
        - VIRTUAL
      - type: string
    Area:
      type: object
      properties:
        tacs:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Tac'
        areaCode:
          $ref: '#/components/schemas/AreaCode'
      oneOf:
      - required:
        - tacs
      - required:
        - areaCode
...
    ServiceAreaRestriction:
      type: object
      properties:
        restrictionType:
          $ref: '#/components/schemas/RestrictionType'
        areas:
          type: array
          items:
            $ref: '#/components/schemas/Area'
        maxNumOfTAs:
          $ref: '#/components/schemas/Uinteger'
        maxNumOfTAsForNotAllowedAreas:
          $ref: '#/components/schemas/Uinteger'
      allOf:
      - oneOf:
        - not:
            required:
            - restrictionType
        - required:
          - areas
      - anyOf:
        - not:
            required:
            - restrictionType
            properties:
              restrictionType:
                type: string
                enum:
                - NOT_ALLOWED_AREAS
        - not:
            required:
            - maxNumOfTAs
      - anyOf:
        - not:
            required:
            - restrictionType
            properties:
              restrictionType:
                type: string
                enum:
                - ALLOWED_AREAS
        - not:
            required:
            - maxNumOfTAsForNotAllowedAreas
...
    CoreNetworkType:
      anyOf:
      - type: string
        enum:
        - 5GC
        - EPC
      - type: string
    RfspIndexRm:
      maximum: 256
      minimum: 1
      type: integer
      nullable: true
    DurationSecRm:
      type: integer
      nullable: true
    OdbPacketServices:
      nullable: true
      anyOf:
      - type: string
        enum:
        - ALL_PACKET_SERVICES
        - ROAMER_ACCESS_HPLMN_AP
        - ROAMER_ACCESS_VPLMN_AP
      - type: string
    Snssai:
      required:
      - sst
      type: object
      properties:
        sst:
          maximum: 255
          minimum: 0
          type: integer
        sd:
          pattern: ^[A-Fa-f0-9]{6}$
          type: string
...
    CommunicationCharacteristics:
      type: object
      properties:
        ppSubsRegTimer:
          $ref: '#/components/schemas/PpSubsRegTimer'
        ppActiveTime:
          $ref: '#/components/schemas/PpActiveTime'
        ppDlPacketCount:
          $ref: '#/components/schemas/PpDlPacketCount'
...
    DateTime:
      type: string
      format: date-time
...
    Supi:
      pattern: ^(imsi-[0-9]{5,15}|nai-.+|.+)$
      type: string
...
    ServiceName:
      anyOf:
      - type: string
        enum:
        - nnrf-nfm
        - nnrf-disc
        - nudm-sdm
        - nudm-uecm
        - nudm-ueau
        - nudm-ee
        - nudm-pp
        - namf-comm
        - namf-evts
        - namf-mt
        - namf-loc
        - nsmf-pdusession
        - nsmf-event-exposure
        - nausf-auth
        - nausf-sorprotection
        - nausf-upuprotection
        - nnef-pfdmanagement
        - npcf-am-policy-control
        - npcf-smpolicycontrol
        - npcf-policyauthorization
        - npcf-bdtpolicycontrol
        - npcf-eventexposure
        - npcf-ue-policy-control
        - nsmsf-sms
        - nnssf-nsselection
        - nnssf-nssaiavailability
        - nudr-dr
        - nlmf-loc
        - n5g-eir-eic
        - nbsf-management
        - nchf-spendinglimitcontrol
        - nchf-convergedcharging
        - nnwdaf-eventssubscription
        - nnwdaf-analyticsinfo
      - type: string
...
    PlmnId:
      required:
      - mcc
      - mnc
      type: object
      properties:
        mcc:
          $ref: '#/components/schemas/Mcc'
        mnc:
          $ref: '#/components/schemas/Mnc'
...